fix: Disable RANSAC plane fitting, segmentation, and removal
fix: Increase min z value for bottom layer, to avoid creating artifacts
on the floor

We did this because table surfaces have already been completely removed
using passthrough filtering. As a consequence, the RANSAC step ends up
removing, at times, planes from objects, like the surface of the book
model.

The downside here is that we can no longer publish table surfaces as
collidable objects.  But for the challenge world, it seems like this
is acceptable, as the objective is merely to correctly label all
objects in the scene